Add units to props documentation
I was not sure if `size` is diameter or radius. Evemtually I found out by trying. Anyway, I guess it's a good idea to have this. Hope you don't mind that I added some info also to `singleColor` which I needed but was not sure how it behaves until I tried out.

This should make the props a bit clearer.

Thanks for this lib, really helpful!